+++ src/components/HarvestingConfiguration/AggregatorInfo/AggregatorContactInfo/AggregatorContactInfo.js
@@ -2,6 +2,7 @@ import _ from 'lodash';
 import React from 'react';
 import PropTypes from 'prop-types';
 import { InfoPopover } from '@folio/stripes/components';
+import { getHeaderWithCredentials } from '@folio/stripes/util';
 
 class AggregatorContactInfo extends React.Component {
   static propTypes = {
@@ -12,14 +13,13 @@ class AggregatorContactInfo extends React.Component {
   constructor(props) {
     super(props);
     this.okapiUrl = props.stripes.okapi.url;
-    this.httpHeaders = Object.assign(
-      {},
-      {
-        'X-Okapi-Tenant': props.stripes.okapi.tenant,
-        'X-Okapi-Token': props.stripes.store.getState().okapi.token,
-        'Content-Type': 'application/json',
-      }
-    );
+
+    this.httpHeaders = {
+      ...getHeaderWithCredentials({
+        tenant: this.props.stripes.okapi.tenant,
+        token: this.props.stripes.store.getState().okapi.token,
+      })
+    };
 
     this.state = {
       contact: null,
@@ -37,9 +37,9 @@ class AggregatorContactInfo extends React.Component {
   }
 
   fechAggregator = (aggregatorId) => {
-    return fetch(`${this.okapiUrl}/aggregator-settings/${aggregatorId}`, {
-      headers: this.httpHeaders,
-    })
+    return fetch(`${this.okapiUrl}/aggregator-settings/${aggregatorId}`,
+      this.httpHeaders,
+    )
       .then((response) => {
         if (!response.ok) {
           return Promise.reject(response);